refactor: 커뮤니티 index.tsx에서 util 함수 분리
